graphics: clear all clippy warnings in the native crate
The graphics native crate is a separate (non-workspace) crate, so the
root `cargo clippy -- -D warnings` gate never covered it — 11 lib lints
+ 5 test lints + 20 deny-by-default `not_unsafe_ptr_arg_deref` on the
raw-ptr ABI accumulated unseen.  Now `cargo clippy --release
--all-targets -- -D warnings` is clean.

Genuine lints fixed:
- redundant `as u32` on `vector_len` (already u32) — 3 sites
- collapsible nested `if` → let-chains (window monitor select, pending
  resize) — edition 2024
- `needless_range_loop` → `iter_mut().enumerate()` for the two 4×4
  matrix-upload buffers
- `key_code < 0 || key_code > 255` → `!(0..=255).contains(&key_code)`
- needless `&` on `File::create(path)`
- `b"...\0"` → `c"..."` C-string literals in headless_safety.rs (5)

Deliberate FFI surface allowed (not "fixed"): the `loft_gl_*` /
`loft_audio_*` entry points are C-ABI fns the runtime + `--native`
codegen call by symbol, taking raw pointers handed over by the runtime.
Marking them `unsafe fn` would change the signature `#[loft_native]`
and the generated callers expect — so a crate-level
`#![allow(clippy::not_unsafe_ptr_arg_deref)]`, sibling to the existing
`missing_safety_doc` allow, with a comment explaining the contract.

Behaviour unchanged (clippy-only refactors; shader-test length 27→26
drops a vestigial counted nul).  headless_safety test green.

+++ lib/graphics/native/src/audio.rs
@@ -235,7 +235,7 @@ pub unsafe extern "C" fn n_audio_play_raw(
     sample_rate: i64,
     volume: f64,
 ) -> i64 {
-    let count = unsafe { store.vector_len(&data) } as u32;
+    let count = unsafe { store.vector_len(&data) };
     let data_ptr = unsafe { store.vector_data_ptr(&data) } as *const f32;
     loft_audio_play_raw(data_ptr, count, sample_rate, volume)
 }
